+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -1628,6 +1628,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
       }
 
       @Test
+      @TestMetadata("GraphPrivateMultibindingMapNotExposedToGrandChild.kt")
+      public void testGraphPrivateMultibindingMapNotExposedToGrandChild() {
+        run("GraphPrivateMultibindingMapNotExposedToGrandChild.kt");
+      }
+
+      @Test
       @TestMetadata("GraphPrivateMultibindingSetNotExposedToChild.kt")
       public void testGraphPrivateMultibindingSetNotExposedToChild() {
         run("GraphPrivateMultibindingSetNotExposedToChild.kt");