+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/symbols/Symbols.kt
@@ -732,6 +732,10 @@ internal class Symbols(
       .single()
   }
 
+  val graphPrivateConstructor by lazy {
+    builtinsFinder.findClass(classIds.graphPrivateAnnotation)!!.constructors.single()
+  }
+
   val assistedConstructor by lazy {
     builtinsFinder
       .findClass(ClassId(metroRuntime.packageFqName, StringNames.ASSISTED.asName()))!!